pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "mcp_finance_agent"
version = "0.1.0"
authors = [
  { name="Andrey Alexeev", email="your.email@example.com" },
]
description = "MCP Finance Agent for portfolio monitoring"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "supabase>=2.3.4",
    "asyncpg>=0.29.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/mcp-finance-agent"
"Bug Tracker" = "https://github.com/yourusername/mcp-finance-agent/issues"

[tool.poetry]
name = "finance-portfolio-agent"
version = "1.0.0"
description = "MCP agent for portfolio analysis and management with Supabase integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src" },
    { include = "mcp_finance_agent" }
]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.9.3"
pydantic = "^2.6.3"
python-dotenv = "^1.0.1"
supabase = "^2.3.4"
asyncpg = "^0.29.0"
tinkoff-investments = "^0.2.0b62"
cryptography = "^42.0.5"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pandas = "^2.2.1"
numpy = "^1.26.4"
pytz = "^2024.1"
grpcio = "^1.62.1"
protobuf = "^4.25.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"
flake8 = "^7.0.0"
pre-commit = "^3.6.0"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing" 